src/classification/requirements.txt
numpy==1.21.0
pandas==1.3.0
SimpleITK==2.0.2
pyradiomics==3.0.1
scikit-learn==0.24.2
xgboost==1.4.2
catboost==0.26
